Split ResultSetHandler, renamed transaction(…), new transaction(…)
Deprecated ResultSetHandler and ResultSetHandlerE in favor of new
ResultSetRunnable, ResultSetRunnableE, ResultSetCallable, and
ResultSetCallableE.  Having both runnable and callable variants allows
implementations to not return anything.  Previously, to satisfy the
interface, something always had to be returned, even if just a
(Void)null.

Renamed Database.executeTransaction(…) to Database.transaction(…).

Added new methods Database.transaction(Runnable) and
Database.transaction(Callable), which are useful when a transaction
scope is needed, but the DatabaseConnection itself is not required.